+++ resolwe_bio/docker_images/dnaseq/packages-manual/samtools.sh
@@ -1,0 +1,21 @@
+#!/bin/bash -e
+
+. /var/cache/build/packages-manual/common.sh
+
+download_and_verify \
+    samtools \
+    samtools \
+    1.6 \
+    ee5cd2c8d158a5969a6db59195ff90923c662000816cc0c41a190b2964dbe49e \
+    https://github.com/samtools/samtools/releases/download/\${version}/samtools-\${version}.tar.bz2 \
+    samtools-\${version}
+
+./configure --without-curses
+make
+
+shopt -s extglob
+rm -rf !(samtools)
+
+add_binary_path \
+    samtools \
+    samtools