+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 * Add Python package ``pandas`` to ``resolwebio/latex`` image
 * Add AF filter to amplicon report
 * Add number of samples to the output of expression aggregator
+* Add primerclip, samtools, picard-tools and bwa to resolwebio/dnaseq docker
 
 Changed
 -------